Refactor outbox publisher flow: Replace leasing with fetching messages
- Updated the publisher flow to use FetchBatchAsync instead of LeaseBatchAsync for message retrieval.
- Modified message selection criteria to include version ceiling filtering and removed row-level locking.
- Adjusted circuit breaker logic to skip messages when open, preventing retry count increments.
- Revised SQL queries in SQL Server and PostgreSQL implementations to reflect the new fetching mechanism.
- Updated documentation to align with changes in the outbox message handling process.
- Enhanced integration tests to ensure proper partition claiming and message processing under various scenarios.

+++ src/Outbox.PostgreSQL/db_scripts/install.sql
@@ -41,7 +41,7 @@ CREATE TABLE IF NOT EXISTS outbox_dead_letter
     created_at_utc       TIMESTAMPTZ(3) NOT NULL,
     retry_count          INT            NOT NULL,
     event_datetime_utc   TIMESTAMPTZ(3) NOT NULL,
-    event_ordinal        SMALLINT       NOT NULL DEFAULT 0,
+    event_ordinal        INT            NOT NULL DEFAULT 0,
     payload_content_type VARCHAR(100)   NOT NULL DEFAULT 'application/json',
     dead_lettered_at_utc TIMESTAMPTZ(3) NOT NULL DEFAULT clock_timestamp(),
     last_error           VARCHAR(2000)  NULL,
